+++ WT4Q/src/app/login/page.tsx
@@ -1,146 +1,11 @@
-'use client';
-import { FC, useState, FormEvent, useTransition } from 'react';
-import { useRouter } from 'next/navigation';
-import styles from './Login.module.css';
-import { API_ROUTES } from '@/lib/api';
+import { cookies } from 'next/headers';
+import { redirect } from 'next/navigation';
+import LoginClient from './LoginClient';
 
-interface LoginRequest {
-  email: string;
-  password: string;
+export default async function LoginPage() {
+  const cookieStore = await cookies();
+  if (cookieStore.get('JwtToken')) {
+    redirect('/');
+  }
+  return <LoginClient />;
 }
-
-const EMAIL_REGEX = /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/;
-
-const Login: FC = () => {
-  const router = useRouter();
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState<string | null>(null);
-  const [emailError, setEmailError] = useState<string | null>(null);
-  const [showPassword, setShowPassword] = useState(false);
-  const [isPending, startTransition] = useTransition();
-  const isLoading = isPending;
-
-  const validate = (): boolean => {
-    let valid = true;
-    if (!EMAIL_REGEX.test(email)) {
-      setEmailError('Please enter a valid email address');
-      valid = false;
-    } else {
-      setEmailError(null);
-    }
-    return valid;
-  };
-
-  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setError(null);
-
-    if (!validate()) return;
-
-    startTransition(async () => {
-      try {
-        const response = await fetch(API_ROUTES.AUTH.LOGIN, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          credentials: 'include',
-          body: JSON.stringify({ email, password } as LoginRequest),
-        });
-        const data: { message?: string } = await response.json();
-
-        if (!response.ok) {
-          throw new Error(data.message || 'Login failed');
-        }
-
-        router.replace('/');
-      } catch (err) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError('An unexpected error occurred');
-        }
-      }
-    });
-  };
-
-  const handleGoogleSignIn = () => {
-    window.location.href = API_ROUTES.GOOGLE_SIGN_IN.AUTH;
-  };
-
-  return (
-    <main className={styles.container}>
-      <section className={styles.card} aria-labelledby="user-login-title">
-        <h1 id="user-login-title" className={styles.title}>
-          User Login
-        </h1>
-        {error && (
-          <div role="alert" className={styles.error}>
-            {error}
-          </div>
-        )}
-        <form onSubmit={handleSubmit} className={styles.form} noValidate>
-          <div className={styles.field}>
-            <label htmlFor="email" className={styles.label}>
-              Email
-            </label>
-            <input
-              id="email"
-              type="email"
-              autoComplete="username"
-              autoFocus
-              required
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              className={styles.input}
-              aria-invalid={!!emailError}
-              aria-describedby={emailError ? 'email-error' : undefined}
-            />
-            {emailError && (
-              <p id="email-error" className={styles.fieldError} role="alert">
-                {emailError}
-              </p>
-            )}
-          </div>
-
-          <div className={styles.field}>
-            <label htmlFor="password" className={styles.label}>
-              Password
-            </label>
-            <div className={styles.passwordWrapper}>
-              <input
-                id="password"
-                type={showPassword ? 'text' : 'password'}
-                autoComplete="current-password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                className={styles.input}
-              />
-              <button
-                type="button"
-                onClick={() => setShowPassword((prev) => !prev)}
-                className={styles.toggleButton}
-                aria-label={showPassword ? 'Hide password' : 'Show password'}
-              >
-                {showPassword ? '🙈' : '👁️'}
-              </button>
-            </div>
-          </div>
-
-          <button type="submit" disabled={isLoading} className={styles.button}>
-            {isLoading ? (
-              <span className={styles.spinner} aria-hidden="true" />
-            ) : (
-              'Sign In'
-            )}
-          </button>
-        </form>
-        <button onClick={handleGoogleSignIn} className={styles.button}>
-          Sign in with Google
-        </button>
-      </section>
-    </main>
-  );
-};
-
-export default Login;
